requirements.txt
# Core Django
Django==4.2.10
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.3.1
django-filter==23.5
django-guardian==2.4.0
django-celery-beat==2.5.0
django-parler==2.3.0
strawberry-graphql==0.219.2
strawberry-graphql-django==0.16.1

# Database
psycopg2-binary==2.9.9
django-postgres-extra==2.0.8
postgis==1.0.4

# Caching & Queue
redis==5.0.1
celery==5.3.6
channels==4.0.0
channels-redis==4.1.0

# Storage & Media
django-storages==1.14.2
pillow==10.1.0
weasyprint==60.2
boto3==1.34.11

# Security
cryptography==41.0.7
django-csp==3.7
django-permissions-policy==4.17.0

# Search & AI
langchain==0.1.1
openai==1.5.0
redis-om==0.2.1
django-vector==0.2.2

# Geospatial
folium==0.15.0
geopy==2.4.1
googlemaps==4.10.0

# Frontend Enhancements
django-htmx==1.17.2
django-compressor==4.4
django-tailwind==3.7.0

# Development & Testing
pytest==7.4.3
pytest-django==4.7.0
factory-boy==3.3.0
coverage==7.3.2
black==23.12.0
isort==5.13.2
mypy==1.7.1
django-debug-toolbar==4.2.0
sentry-sdk==1.38.0
django-extensions==3.2.3 